decision_making/decision_making.py
import numpy as np

class DecisionMaking:
    def __init__(self):
        pass

    def evaluate_strategies(self, data):
        """
        Evaluate different decision-making strategies based on the provided data.

        Parameters:
        - data: numpy array or pandas DataFrame
            The data to be used for evaluating strategies.

        Returns:
        - dict
            A dictionary containing the evaluation results for each strategy.
        """
        # Placeholder for strategy evaluation logic
        strategies = ['strategy_1', 'strategy_2', 'strategy_3']
        results = {strategy: np.random.rand() for strategy in strategies}
        return results

    def assess_risks(self, data):
        """
        Assess risks associated with different decisions based on the provided data.

        Parameters:
        - data: numpy array or pandas DataFrame
            The data to be used for risk assessment.

        Returns:
        - dict
            A dictionary containing the risk assessment results for each decision.
        """
        # Placeholder for risk assessment logic
        decisions = ['decision_1', 'decision_2', 'decision_3']
        risks = {decision: np.random.rand() for decision in decisions}
        return risks

    def predict_outcomes(self, data):
        """
        Predict outcomes of different decisions based on the provided data.

        Parameters:
        - data: numpy array or pandas DataFrame
            The data to be used for outcome prediction.

        Returns:
        - dict
            A dictionary containing the predicted outcomes for each decision.
        """
        # Placeholder for outcome prediction logic
        decisions = ['decision_1', 'decision_2', 'decision_3']
        outcomes = {decision: np.random.rand() for decision in decisions}
        return outcomes
